apply xenon c634a7f0d00944ca901ad6a4eb4b0d0a4e8e41a
https://github.com/vmware/xenon/commit/7c634a7f0d00944ca901ad6a4eb4b0d0a4e8e41a

+++ dcs-core/src/main/java/com/dcentralized/core/common/NodeSelectorState.java
@@ -13,12 +13,10 @@
 
 package com.dcentralized.core.common;
 
-
 import java.util.EnumSet;
 
 import com.dcentralized.core.services.common.NodeGroupService.NodeGroupState;
 import com.dcentralized.core.services.common.NodeGroupUtils;
-
 
 public class NodeSelectorState extends ServiceDocument {
 
@@ -55,10 +53,36 @@ public class NodeSelectorState extends ServiceDocument {
         return AVAILABLE.contains(calculateStatus(host, groupState));
     }
 
+    /**
+     * Path of this node selector's nodegroup
+     */
     public String nodeGroupLink;
-    public Long replicationFactor;
-    public int replicationQuorum;
+
+    /**
+     * Determines how many nodes need to receive a copy of a change.
+     * Null means: all nodes are replication targets.
+     */
+    public Integer replicationFactor;
+
+    /**
+     * Determines how many nodes need to accept a change for the change to succeed.
+     * Null means: use membershipQuorum as the replication quorum.
+     */
+    public Integer replicationQuorum;
+
+    /**
+     * Determines how many nodes need to be alive and reachable for the nodegroup
+     * to be considered 'available'
+     */
     public int membershipQuorum;
+
+    /**
+     * Nodegroup membership update time
+     */
     public long membershipUpdateTimeMicros;
+
+    /**
+     * Nodegroup status
+     */
     public Status status;
 }
